+++ frontend/src/components/product/product.component.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { withRouter, Redirect } from 'react-router-dom';
+import { withRouter } from 'react-router-dom';
 import Button from '@material-ui/core/Button';
 import { getProduct } from '../../services/Products';
 import './product.styes.scss';
